requirements.txt
# 设置插件依赖的版本
pytest==3.7.3
requests==2.19.1
pytest-allure-adaptor==1.7.10